--- Snake.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Snake", "Snake\Snake.csproj", "{C7BC954A-8473-4964-98E0-776A4987C7D0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C7BC954A-8473-4964-98E0-776A4987C7D0}.Debug|x86.ActiveCfg = Debug|x86
-		{C7BC954A-8473-4964-98E0-776A4987C7D0}.Debug|x86.Build.0 = Debug|x86
-		{C7BC954A-8473-4964-98E0-776A4987C7D0}.Release|x86.ActiveCfg = Release|x86
-		{C7BC954A-8473-4964-98E0-776A4987C7D0}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
