--- Snake/MainWindow.xaml
@@ -1,18 +1,0 @@
-﻿<Window x:Class="Snake.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="MainWindow" Height="681.469" Width="920.655" KeyDown="Window_KeyDown">
-    <Grid Visibility="Visible">
-        <Button Content="Start" Height="23" HorizontalAlignment="Left" Margin="12,12,0,0" Name="button1" VerticalAlignment="Top" Width="75" Click="button1_Click" />
-
-        <Label Content="Score" Height="49" HorizontalAlignment="Left" Margin="718,0,0,0" Name="label1" VerticalAlignment="Top" FontFamily="Showcard Gothic" FontSize="30" Width="124" />
-        <Label Content="0" Height="50" HorizontalAlignment="Left" Margin="718,54,0,0" Name="lblScore" VerticalAlignment="Top" Width="158" FontFamily="Showcard Gothic" FontSize="30" />
-        <Canvas Height="600" HorizontalAlignment="Left" Margin="111,12,0,0" Name="canvas1" VerticalAlignment="Top" Width="600">
-            <Image Height="600" Width="600" Source="Resources/logo.png"/>
-
-        </Canvas>
-        <TextBlock Height="300" Name="tbGameOver" Text="   Game Over!" Width="300" FontFamily="Levenim MT" FontSize="36" Background="#79483232" Foreground="#FFFC0000" Visibility="Hidden" Margin="111,12,164,12" />
-        <ProgressBar HorizontalAlignment="Left" Height="24" Margin="111,617,0,0" VerticalAlignment="Top" Width="498" Value="60"/>
-        <Label x:Name="label" Content="х 4" HorizontalAlignment="Left" Margin="614,610,0,0" VerticalAlignment="Top" Width="97" FontFamily="LCDMono2" FontSize="30"/>
-    </Grid>
-</Window>
