+++ Snake2/MainWindow.xaml
@@ -1,0 +1,29 @@
+﻿<Window x:Class="Snake2.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:Snake2"
+        mc:Ignorable="d"
+        Title="MainWindow" Height="645" Width="900" KeyDown="Window_KeyDown" ResizeMode="NoResize">
+    <Grid Visibility="Visible" Height="Auto" Width="Auto">
+        <Button Content="Start" FontFamily="Levenim MT" FontSize="100"  Height="Auto" HorizontalAlignment="Center" Name="button1" VerticalAlignment="Center" Width="Auto" Click="button1_Click" />
+        <Label Content="Score" Visibility="Hidden" Height="49" HorizontalAlignment="Left" Margin="718,0,0,0" Name="label1" VerticalAlignment="Top" FontFamily="Showcard Gothic" FontSize="30" Width="124" />
+        <Label Content="0" Visibility="Hidden" Height="50" HorizontalAlignment="Left" Margin="718,54,0,0" Name="lblScore" VerticalAlignment="Top" Width="158" FontFamily="Showcard Gothic" FontSize="30" />
+        <Canvas Height="600" HorizontalAlignment="Left" Margin="111,12,0,0" Name="canvas1" VerticalAlignment="Top" Width="600" Visibility="Hidden">
+            <Image Height="600" Width="600" Source="Resources/logo.png"/>
+        </Canvas>
+        <TextBlock Name="tbGameOver" TextWrapping="Wrap" Text="    Game Over!Press R to restart!" Width="300" FontFamily="Levenim MT" FontSize="36" Background="#79483232" Foreground="#FFFC0000" Visibility="Hidden" Margin="279,186,315,325" />
+        <ProgressBar x:Name="progressBar" HorizontalAlignment="Left" Height="24" Margin="485,347,-89,0" VerticalAlignment="Top" Width="498" Value="100" RenderTransformOrigin="0.5,0.5" Visibility="Hidden">
+            <ProgressBar.RenderTransform>
+                <TransformGroup>
+                    <ScaleTransform/>
+                    <SkewTransform/>
+                    <RotateTransform Angle="90"/>
+                    <TranslateTransform/>
+                </TransformGroup>
+            </ProgressBar.RenderTransform>
+        </ProgressBar>
+        <Label x:Name="Score_Multiplier" Visibility="Hidden" Content="х 1" HorizontalAlignment="Left" Margin="759,109,0,0" VerticalAlignment="Top" Width="97" FontFamily="LCDMono2" FontSize="30" RenderTransformOrigin="2.485,-6.784"/>
+    </Grid>
+</Window>
