+++ Snake2/Resources/fruit.png
