+++ Snake2/Snake2.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30717.126
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Snake2", "Snake2.csproj", "{2FB1B1A8-48B8-472F-9CF3-707AB6DB6725}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2FB1B1A8-48B8-472F-9CF3-707AB6DB6725}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2FB1B1A8-48B8-472F-9CF3-707AB6DB6725}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2FB1B1A8-48B8-472F-9CF3-707AB6DB6725}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2FB1B1A8-48B8-472F-9CF3-707AB6DB6725}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1A5B7091-79CE-4E3F-B17C-E92BB7C0292F}
+	EndGlobalSection
+EndGlobal
